Fix compiling error by using long long instead of float
C++ is kind of confusing when you write the code for the first time, but if you've learned haxe then yeah

+++ project/src/media/AudioBuffer.cpp
@@ -6,7 +6,7 @@ namespace lime {
 
 	static int id_bitsPerSample;
 	static int id_channels;
-	static int id_data;
+	static long long id_data;
 	static int id_sampleRate;
 	static bool init = false;
